feat: Set up vehicle detection pipeline
As a first step in creating the vehicle detection pipeline, all the 
relevant functions from the lessons were organized to separate the image 
processing from the classification.

The Detector class handles searching through the images and extracting 
features to pass to the classifier. This is where the search area and 
feature extraction parameters will be tuned to improve detection.

The Model class stores and trains the classification model. This 
abstracts the classification from the image processing and creates a 
wrapper around prediction and training functions.

Any extra functions are in helper_functions.py. This is for smaller 
functions, functions that are used in multiple places, or functions that 
don't belong to a specific task such as "detection" or "classification".