feat: Search windows and return prediction boxes
The Detector class can now create an array of sliding windows, extract 
features from those windows, and make predictions based on those 
featuers.

The find_cars() function was also changed to return a list of bounding 
boxes instead of an image. This change was made so find_cars() and 
search_windows() have the same API and can be interchangeable.