feat: Load project video into pipeline
The pipeline now uses the project video instead of test images. This is 
both to test the overall pipeline on a video stream, and also to utilize 
multi-frame heatmap smoothing that does not work on single images.

Converting images from jpg format [0,255] to png format [0.0,1.0] was 
also moved into the find_cars function because a png format image was 
not showing up in the output video file.