feat: Create generators for data
Previously, only a small fraction of the available data was being used. 
Now with the generators, all of the data loading can be handled in one 
location and only the images currently being processed are stored in 
memory.

This does mean that training the model will take much longer, but it 
only needs to be trained once and should lead to more accurate results.