feat: Get pipeline to stable state
This commit is very large and covers a lot of things, which is because I 
ran into some git issues that caused me to have to roll back and copy 
all my changes over. This should solve the git issues but results in an 
unwieldy commit with a very broad scope.

The main focus of this commit was getting the pipeline to a stable state 
that does not a completely terrible job of detecting vehicles. 
Previously, it would run but nothing was detected. Now things have been 
built back up and the pipeline at least produces predictions.