chore: Create IPython notebook to test individual pipeline components
This is where the bulk of the work I have done since my last submission was done. I was still struggling a lot with trying to figure out where my pipeline was failing, so I worked alongside the Q&A video that was put out by Udacity and I rebuilt the pipeline function by function, testing each one individually, in this notebook.

There are examples showing each function working correctly, and everything working together so the parameters could be tweaked and performance could be improved.